==> Scripts/1. Data Wrangling Orders.csv.ipynb <==
# Data Wrangling Orders.CVS

# Importing libraries and Data sets

# Wrangling Orders

## Cross Tab

## Rename columns

## Updating data types

# Creating a Data Dictionary

## Transpose Data Dictionary 

## Set Data Dictionary

# Creating Profile for User 1

# Save Datasets

------------------

# Importing libraries and Data sets

#Importing Libraries
import pandas as pd
import numpy as np
import os

#creating DF pathway
path=(r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis\Data')

#Import dataframe orders
df_ord= pd.read_csv(os.path.join(path,'original data','orders.csv'), index_col=False)

#import dataframe products
df_prod = pd.read_csv(os.path.join(path,'original data','product.csv'), index_col=False)

#import df department ID
df_dept = pd.read_csv(os.path.join(path,'original data','departments.csv'), index_col=False)

# Wrangling Orders Dataset

#Dropping Eval column
df_ord= df_ord.drop(columns = ['eval_set'])

#value count days_since_prior_order
df_ord['days_since_prior_order'].value_counts(dropna = False)

## Cross Tabs 

#CReating crosstab for for days_since_prior_order
crosstab = pd.crosstab(df_ord['days_since_prior_order'], df_ord['order_number'], dropna = False)

#creating crosstab for excel spreedsheet to save to client files
crosstab.to_clipboard()

## Rename Columns

#rename dow colmn to days of week
df_ord.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

#column check
df_ord.head()

#redefine order_id as string
df_ord['order_id'] = df_ord['order_id'].astype('str')

#updating order_number to clients_order_number column name
df_ord.rename(columns={'order_number':'clients_order_number'}, inplace=True)

## Updating data types

#Change different variable in the df_ords df to a different,more approepirate data type [user_id tp string]
df_ord['user_id']=df_ord['user_id'].astype('string')

#checking user id to confirm data type has updated to string 
df_ord['user_id'].dtype

# Create Data Dictionary

## Transpose Data Dictionary 

df_dept.head()

#Tranposing Departments.cvs
df_dept.T

#rename dept df
df_dept_t = df_dept.T

#reset DF index dept
df_dept_t.reset_index()

#create new header
new_header = df_dept_t.iloc[0]

#create new dpt_df w/out 1st row
df_dep_t_new = df_dept_t[1:]

#attaching header to new dpt_df
df_dep_t_new.columns = new_header

df_dep_t_new.head()

## Set Data Dictionary

#creating data dict
data_dict=df_dep_t_new.to_dict('index')

# Create profile for user 1

#Profile on User_id 1
df_ord_user1=df_ord.loc[df_ord['user_id']=='1']

#User_id 1 Behaviors [Includes all data from orders.csv]
df_ord_user1

# Save datasets

#Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data”
df_ord.to_csv(os.path.join(path,'Prepared Data', 'orders_wrangled.csv'), index=False)

#Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data”
df_dep_t_new.to_csv(os.path.join(path,'Prepared Data', 'departments_wrangled.csv'), index=False)

#Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data”
df_dep_t_new.to_csv(os.path.join(path,'Prepared Data', 'departments_wrangled.csv'))
