==> Scripts/2. Data Consistency Checks Products.csv and Orders.csv.ipynb <==
# Data Consistency Checks Products and orders

# Importing Libraries and datasets

# Consistency Checks on Products

## Checking for Missing Values

## Checking for Duplicates

# Consistency Check on Orders

# Export Data sets

--------

# Importing Libraries and datasets

#Import Libraries
import pandas as pd
import numpy as np
import os

#creating pathway for dataframes
path=os.path.join(r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis\Data')

#import products.csv
df_prods= pd.read_csv(os.path.join(path,'Original Data', 'product.csv'),index_col=False)

#import orders_wrangled.csv
df_ords= pd.read_csv(os.path.join(path,'Prepared Data', 'orders_wrangled.csv'))

# Consistency Checks on Products

# Describe check on products df
df_prods.describe()

#check for mixed types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

## Checking for Missing Values

#checking for missing data
df_prods.isnull().sum()

#new df for product Nan values
df_nan=df_prods[df_prods['product_name'].isnull()==True]

#check
df_nan

#new products df with no missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

#shape check
df_prods_clean.shape

## Checking for Duplicates

#new df for finding duplicates
df_dups=df_prods_clean[df_prods_clean.duplicated()]

#check
df_dups

df_prods_clean.shape

#new df with no duplicates
df_prods_clean_no_dups=df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# Consistency Check on Orders

#adding code to get rid of exponents
pd.options.display.float_format = '{:.2f}'.format

#exploritory on Orders DF
df_ords.describe()

# Checking for mixed type data - no issues found
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

# Missing Values check
df_ords.isnull().sum()

# Export Data sets

#exporting cleaned products df
df_prods_clean_no_dups.to_csv(os.path.join(path,'Prepared Data', 'products_checked.csv'), index=False)

# Saving orders.csv
df_ords.to_csv(os.path.join(path,'Prepared Data', 'orders_checked.csv'), index=False)
