==> Scripts/3. Joining orders_prods_combined.ipynb <==
# Importing Libraries and DFs

# Checking DF Dimensions

# Merging orders_prior and orders Dataframes

# Checking DF and Saving

# Exporting Datasets

-------

# Importing Libraries and DFs

#Import Libraries
import pandas as pd
import numpy as np
import os

#create df pathway
path= (r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis\Data')

#Importing orders_prior DF
df_ords=pd.read_csv(os.path.join(path,'prepared data', 'orders_checked.csv'), index_col= False) 

#Importing products_prior DF
df_prods=pd.read_csv(os.path.join(path,'prepared data', 'products_checked.csv'), index_col= False) 

#Importing products_prior DF
df_ords_prior=pd.read_csv(os.path.join(path,'original data', 'orders_products_prior.csv'), index_col= False) 

# Checking DF Dimensions

df_ords_prior.head()

df_ords_prior.shape

df_ords.head()

df_ords.shape

# Merging orders_prior and orders Dataframes

#merging order df and order_prior df
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

# Checking DF and Saving

#checking value counts in df-merge
df_merged_large['_merge'].value_counts()

df_merged_large.shape

#removing _merge column for upcoming merge
df_merged_larges=df_merged_large.drop(columns=['_merge'])

# Exporting Datasets

# Export data to pkl
df_merged_large.to_pickle(os.path.join(path,'Prepared Data', 'orders_products_combined.pkl'))
