==> Scripts/4. Joining order_products_merged.ipynb <==
# Importing Libraries and DFs

# Checking Dimensions

# Combining Tables

# Quality check on merged table 

# Exporting Datasets

---------------

# Importing Libraries and DFs

#Import Libraries
import pandas as pd
import numpy as np
import os

#create df pathway
path= (r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis\Data')

#Importing products_prior DF
df_prods=pd.read_csv(os.path.join(path,'prepared data', 'products_checked.csv'), index_col= False) 

# import orders_products_combined
df_ords_prods=pd.read_pickle(os.path.join(path,'Prepared Data', 'orders_products_combined.pkl'))

# Checking Dimensions

# shape of ords_prods
df_ords_prods.shape

#Task5 analysis of DF to determine merging key
df_prods.head()

df_ords_prods.head()

df_prods.shape

df_ords_prods.shape

# Combining Tables

# merging prods df with ords_prods DF
df_orders_products_merged = df_ords_prods.merge(df_prods, on = 'product_id',  indicator = True)

# Quality check on merged table 

df_orders_products_merged.head()

# checking the merge flag
df_orders_products_merged['_merge'].value_counts()

#dropping merge column
df_orders_products_merged=df_orders_products_merged.drop(columns = ['_merge'])

df_orders_products_merged.shape

# Exporting Datasets

#Save new merge df at pkl
df_orders_products_merged.to_pickle(os.path.join(path,'Prepared Data', 'orders_products_merged.pkl'))
