==> Scripts/5.  Deriving New Variables Orders_prods_merged.ipynb <==
# 4.7 Deriving New Variables Orders_prods_merged

# Importing Libraries and Dataframes

# Sub-set practice

## Price label flag(User syntax)

## Price_range Flag (loc)

# Deriving Variables Full dataset

## Price_range (loc)

## Busiest_day flag (Loop)

# Derving Stakeholder Questions

## Busiest_days flag

## Busiest_period_of_day flag

# Export Dataset

------------

# Importing Libraries and Dataframes

import pandas as pd
import numpy as np
import os

#pathway to datasets
path =(r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis\Data')

df_ords_prods_merge=pd.read_pickle(os.path.join(path,'Prepared Data','orders_products_merged.pkl'))

# checking dimensions
df_ords_prods_merge.head()

df_ords_prods_merge.shape

# Sub-set practice

#creating a subset
df=df_ords_prods_merge[:1000000]

## Price label flag(User syntax)

#user def- price label syntax
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return np.nan

#creating flag for user syntax price label
df['price_range'] = df.apply(price_label, axis=1)

#value count of Price_range column
df['price_range'].value_counts(dropna=False)

#most expensive product within the subset
df['prices'].max()

## Price_range Flag (loc)

#setting loc for high range products
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

#setting loc for mid range products
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

#setting loc for low range products
df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna=False)

# Deriving Variables Full dataset

## Price_range (loc)

#setting loc for high range products
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

#setting loc for mid range products
df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

#setting loc for low range products
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merge['price_range_loc'].value_counts(dropna=False)

## Busiest_day flag (Loop)

#Frequency count  DOW 
df_ords_prods_merge['orders_day_of_week'].value_counts(dropna=False)

#creating if loop to group busiest_day
result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("busiest day")
  elif value == 4:
    result.append("least busy")
  else:
    result.append("regularly busy")

df_ords_prods_merge['busiest_day'] = result

#busiest_day results
df_ords_prods_merge['busiest_day'].value_counts(dropna=False)

--------

# Derving Stakeholder Questions

## Busiest_days flag

#Redefining busiest_day to busiest_days
result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value <= 1:
    result.append("busiests days")
  elif value == 3:
    result.append("least busy days")
  elif value == 4:
    result.append("least busy days")
  else:
    result.append("regularly busy days")

# set results to busiest_days
df_ords_prods_merge['busiest_days'] = result

#Frequency count busiest order hours
df_ords_prods_merge['busiest_days'].value_counts(dropna=False)

## Busiest_period_of_day flag

#Frequency count busiest order hours
df_ords_prods_merge['order_hour_of_day'].value_counts(dropna=False)

#creating subset for busiest_period_of_day (L-12 to 9 M- 10 to 16 A- 17 to 23)
result = []

for value in df_ords_prods_merge["order_hour_of_day"]:
  if value <= 9:
    result.append("least orders")
  elif value >= 17:
    result.append("average orders")
  else:
    result.append("most orders")

# set results to busiest_days
df_ords_prods_merge['busiest_period_of_day'] = result

#Value count busiest order hours
df_ords_prods_merge['busiest_period_of_day'].value_counts(dropna=False)

# Export Dataset

#Export df_ords_prods_merge into a pkl
df_ords_prods_merge.to_pickle(os.path.join(path,'Prepared Data', 'orders_products_merge.pkl'))
