==> Scripts/6. Grouping and Aggrigating Data Reading ords_prods_merged.ipynb <==
# 4.8.1 Grouping and Aggrigating Data Reading ords_prods_merged

# Importing Libraries and Dataframes

# Subset and Groupby Exercise

### Grouping Data with pandas

# Creating a Loyalty flag

# Export Data

--------------

# Importing Libraries and Dataframes

import pandas as pd
import numpy as np
import os

#pathway to datasets
path =(r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis\Data')

#Importing dataset
ords_prods_merge=pd.read_pickle(os.path.join(path,'Prepared Data','orders_products_merged.pkl'))

#creating PD setting, display max columns 100 and max rows 100
pd.set_option('display.max_rows', 100, 'display.max_columns', 100,)

#turning off max rows pandas setting
pd.options.display.max_rows = None

# Subset and Groupby Exercise

#creating 1 mil row sub-set
df=ords_prods_merge[:1000000]

df.shape

df.head(10)

### Grouping Data with pandas

#grouping product_name
df.groupby('product_name')

#agg mean of order number
df.groupby('department_id').agg({'clients_order_number': ['mean']})

#calculating mean
df.groupby('department_id')['clients_order_number'].mean()

#calculating mean, and min max
df.groupby('department_id').agg({'clients_order_number': ['mean', 'min', 'max']})

# Creating a Loyalty flag

#creating max order column
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['clients_order_number'].transform(np.max)

ords_prods_merge.head(100)

#Loyal Customer LOC tag
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

#value count loyalty customer
ords_prods_merge['loyalty_flag'].value_counts(dropna=False)

ords_prods_merge[['user_id', 'loyalty_flag', 'clients_order_number']].head(60)

# Export Data

#Export df_ords_prods_merge to PKL
ords_prods_merge.to_pickle(os.path.join(path,'Prepared Data', 'orders_products_merged.pkl'))
