==> Scripts/8. Preparing Customer.csv.ipynb <==
# 4.9 Preparing Customer.CSV

# Importing Libraries and Dataset

#  Wrangling Customer.CSV

## Addressing Columns (rename and drop)

## Changing Data Types

## Missing Values

## Duplicates

# Merging Customers with ords_prods

## Merging Tables

# Export Data

----------------

# Importing Libraries and Dataset

#Importing Libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# Creating pathway to datasets
path =(r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis')

# Importing dataset customer
cust=pd.read_csv(os.path.join(path,'data','original Data','customers.csv'))

# Importing dataset ords_prods
ords_prods=pd.read_pickle(os.path.join(path,'data','prepared Data','orders_products_merged.pkl'))

#  Wrangling Customers.csv

#Checking Columns
cust.head(10)

#Row count
cust.count

# data description
cust.describe()

#checking types
cust.info()

#fam_status value_counts
cust['fam_status'].value_counts(dropna=False)

#date_joined value_counts
cust['date_joined'].value_counts(dropna=False)

## Addressing Columns (rename and drop)

# dropping colmn surname for privacy
cust= cust.drop(columns =['Surnam'])

#ranameing column STATE to state_usa for improved identification
cust.rename(columns= {'STATE':'state_usa'}, inplace= True)

# dropping colmn first name 
cust= cust.drop(columns =['First Name'])

## Changing Data Types

# changing user_id from int64 to string
cust['user_id'] = cust['user_id'].astype('str')

# changing user_id from int64 to string on ords_prods
ords_prods['user_id'] = ords_prods['user_id'].astype('str')

# changing order_id from int64 to string 
ords_prods['order_id'] = ords_prods['order_id'].astype('str')

## Missing Values

#missing value count
cust.isnull().sum()

# creating df for NaN values
df_fn = cust[cust['First Name'].isnull()==True]

#checking NaN values
df_fn

## Duplicates

#creating subset for duplicates
df_dups = cust[cust.duplicated()]

#no duplicates found
df_dups.info()

# Merging Customers with ords_prods

# dimension check
cust

ords_prods.info()

## Merging Tables

#merging tables on orders_prods using 'inner' statement
orders_products_all = ords_prods.merge(cust, on=['user_id'], indicator=True)

orders_products_all.head()

orders_products_all.columns

### Quality checking results

#Value counts
orders_products_all['_merge'].value_counts(dropna=False)

orders_products_all_outer = ords_prods.merge(cust, on=['user_id'], how='outer', indicator=True)

#Value counts outer
orders_products_all_outer['_merge'].value_counts(dropna=False)

#  Export Data

### Cleaning up before Export

#drop _merge column
orders_products_all = orders_products_all.drop(columns=['_merge'])

#rearrange columns  for better flow
orders_products_all=orders_products_all[['order_id','user_id','clients_order_number','orders_day_of_week','order_hour_of_day','days_since_prior_order','product_id','add_to_cart_order','reordered','product_name','aisle_id','department_id','prices','Gender','state_usa','Age','date_joined','n_dependants','fam_status','income','price_range_loc','busiest_period_of_day','busiest_days','loyalty_flag','spending_flag','FRQ_prior_order_flag']]

orders_products_all.head()

### Exporting

orders_products_all.to_pickle(os.path.join(path, 'Data','Prepared Data', 'orders_products_all.pkl'))
