==> Scripts/10. ord_prods_all final analysis Part1.ipynb <==
# 10.1 Final Analysis Questions (Answers Tasks 1-4)

# Import Libraries and Datasets

## Set PD code

# Reviewing for PHI/Security Implications

# Analysis of Regional Sales

## Creating a Regional Flag

##  Creating a Crosstab for regional and spending flag for analysis

# Creating new data Set without Low-Activity Customers

## Create Exclusionary Flag for Low-Spenders

## Exporting Active members data set as orders_products_active.pkl

------

# Import Libraries and Datasets

#Import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

#pathway to datasets
path =(r'C:\Users\colmu\Desktop\Career Foundry\Achievement 4\Instacart Basket Analysis')

#Importing dataset
ords_prods=pd.read_pickle(os.path.join(path, 'data','Prepared Data','orders_products_all.pkl'))

### Set PD code

#creating PD setting, display max columns
pd.set_option('display.max_columns', 1000)

# Reviewing for PHI/Security Implications

#reviewing column headers and first 5 rows for PHI implications. PHI previously removed, no changes made.
ords_prods.head()

# Analysis of Regional Sales

## Creating a Regional Flag

#creating LOC for Northeast region
ords_prods.loc[ords_prods['state_usa'].isin(['Maine','New Hampshire','Vermont','Massachusetts', 'Rhode Island','Connecticut','New York','Pennsylvania','New Jersey']), 'regional_flag'] = 'Northeast'

#creating LOC for Midwest
ords_prods.loc[ords_prods['state_usa'].isin(['Wisconsin','Michigan','Illinois','Indiana', 'Ohio', 'North Dakota','South Dakota''Nebraska','Kansas''Minnesota','Iowa''Missouri']), 'regional_flag'] = 'Midwest'

#creating LOC for South region
ords_prods.loc[ords_prods['state_usa'].isin(['Delaware', 'Maryland','District of Columbia''Virginia''West Virginia','North Carolina','South Carolina', 'Georgia','Florida',' Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana']), 'regional_flag'] = 'South'

#creating LOC for West region
ords_prods.loc[ords_prods['state_usa'].isin(['Idaho', 'Montana','Wyoming','Nevada','Utah','Colorado','Arizona','New Mexico','Alaska','Washington','Oregon','California','Hawaii']), 'regional_flag'] = 'West'

#checking regional Flag value counts
ords_prods['regional_flag'].value_counts(dropna=False)

#Total Row count for Regional_flag
ords_prods['regional_flag'].count()

##  Creating a Crosstab for regional and spending flag for analysis

#creating a crosstab 
region_crosstab=pd.crosstab(ords_prods['regional_flag'],ords_prods['spending_flag'],dropna=False)

#check
region_crosstab

# Creating new data Set without Low-Activity Customers

## Create Exclusionary Flag for Low-Spenders

#creating max column for client orders
ords_prods['max_order'] = ords_prods.groupby(['user_id'])['clients_order_number'].transform(np.max)

#Creating Exclusionary row for low acitvity and active cusotmers
result = []

for value in ords_prods["max_order"]:
  if value < 5:
    result.append("Low_activity")
  else:
    result.append("Active")

# set results to Customer_activity
ords_prods['Customer_activity'] = result

ords_prods['Customer_activity'].value_counts()

#creating subset of low_activity customers
orders_products_Low_Activity = ords_prods[ords_prods['Customer_activity'] =='Low_activity']

#creating subset of active customers
orders_products_active = ords_prods[ords_prods['Customer_activity'] =='Active']

#check
orders_products_Low_Activity.shape

#check
orders_products_active.shape

#matching total data set row count
1440295 + 30964564

## Exporting Active members data set as orders_products_active.pkl

#Export orders_products_active as PKL
orders_products_active.to_pickle(os.path.join(path,'Data','Prepared Data','orders_products_active.pkl'))
